pyproject.toml
[project]
name = "langgraph-research-assistant"
version = "0.1.0"
dependencies = [
  "langgraph",
  "langgraph-cli[inmem]",
  "langchain",
  "langchain-openai",
  "openai",
  "python-dotenv"
]